refactor(types): improve types and type guards in `wrapper.ts`
The following changes concern the  `isSignal`, `isComputed`, `isWatcher` functions:

- switched to standard function syntax instead of using `let` (now the actual brand checking logic is stored in an internal object)
- upgraded to bona fide type guards with `_ is T` return types
- removed the `any` parameter types
- added type parameters to  `isSignal` and `isComputed`
- wrapped the arguments with `Object` to prevent the `in` operator from throwing when it receives a primitive value

I also opted to switch the empty `#brand` methods to properties set to `true` in the `State`, `Computed`, and `Watcher` classes. This was solely done for aesthetics; for some reason those empty methods bothered me. If this is not desired, please let me know and I'll gladly revert those bits 😅